split test_other_cases
