+++ tests/drum/constants.py
@@ -41,6 +41,7 @@ XGB = "xgboost"
 KERAS = "keras"
 KERAS_JOBLIB = "keras_joblib"
 SKLEARN = "sklearn"
+SKLEARN_NO_ARTIFACTS = "sklearn_no_artifacts"
 SKLEARN_BINARY = "sklearn_binary"
 SKLEARN_REGRESSION = "sklearn_regression"
 SKLEARN_MULTICLASS = "sklearn_multiclass"
@@ -160,6 +161,7 @@ R_INT_COLNAMES_BINARY = "R_int_colnames_binary"
 R_INT_COLNAMES_MULTICLASS = "R_int_colnames_multiclass"
 R_FIT = "R_fit"
 R_PREDICT_SPARSE = "R_predict_sparse"
+R_NO_ARTIFACTS = "r_no_artifacts"
 JAVA = "java"
 JULIA = "julia"
 PYTHON_UNSTRUCTURED = "python_unstructured"
