+++ tests/drum/run_integration_tests_in_framework_container.sh
@@ -22,7 +22,10 @@ echo "Running pytest:"
 
 pip install pytest pytest-xdist
 
-pytest tests/drum/test_inference_per_framework.py tests/drum/test_fit_per_framework.py --framework-env $1 \
+pytest tests/drum/test_inference_per_framework.py \
+       tests/drum/test_fit_per_framework.py \
+       tests/drum/test_other_cases_framework.py \
+       --framework-env $1 \
        --junit-xml="$GIT_ROOT/results_integration.xml" \
        -n auto
