+++ tests/drum/test_other_cases.py
@@ -9,16 +9,15 @@ from uuid import uuid4
 import pandas as pd
 import pytest
 import requests
-from unittest.mock import Mock, patch, DEFAULT
+from unittest.mock import patch, DEFAULT
 import yaml
 
-from datarobot_drum.drum.enum import CUSTOM_FILE_NAME, CustomHooks, ArgumentsOptions
+from datarobot_drum.drum.enum import CUSTOM_FILE_NAME, ArgumentsOptions
 
 from datarobot_drum.resource.drum_server_utils import DrumServerRun
 from datarobot_drum.resource.utils import (
     _exec_shell_cmd,
     _create_custom_model_dir,
-    _cmd_add_class_labels,
 )
 
 from .constants import (
@@ -28,21 +27,11 @@ from .constants import (
     CODEGEN_AND_SKLEARN,
     REGRESSION,
     REGRESSION_INFERENCE,
-    UNSTRUCTURED,
     BINARY,
     PYTHON,
     NO_CUSTOM,
-    PYTHON_ALL_PREDICT_STRUCTURED_HOOKS,
-    PYTHON_ALL_PREDICT_UNSTRUCTURED_HOOKS,
     R,
-    R_ALL_PREDICT_STRUCTURED_HOOKS,
-    R_ALL_PREDICT_STRUCTURED_HOOKS_LOWERCASE_R,
-    R_ALL_PREDICT_UNSTRUCTURED_HOOKS,
-    R_ALL_PREDICT_UNSTRUCTURED_HOOKS_LOWERCASE_R,
     DOCKER_PYTHON_SKLEARN,
-    R_INT_COLNAMES_BINARY,
-    R_INT_COLNAMES_MULTICLASS,
-    MULTICLASS,
     TESTS_DATA_PATH,
 )
 from datarobot_drum.drum.enum import MODEL_CONFIG_FILENAME
@@ -60,36 +49,6 @@ from datarobot_drum.drum.enum import ExitCodes
 
 
 class TestOtherCases:
-    @pytest.mark.parametrize(
-        "framework, problem, language", [(SKLEARN, BINARY, PYTHON), (RDS, BINARY, R)]
-    )
-    def test_bin_models_with_wrong_labels(
-        self, resources, framework, problem, language, tmp_path,
-    ):
-        custom_model_dir = _create_custom_model_dir(
-            resources, tmp_path, framework, problem, language,
-        )
-
-        input_dataset = resources.datasets(framework, problem)
-        cmd = "{} score --code-dir {} --input {} --target-type {}".format(
-            ArgumentsOptions.MAIN_COMMAND,
-            custom_model_dir,
-            input_dataset,
-            resources.target_types(problem),
-        )
-        if problem == BINARY:
-            cmd = cmd + " --positive-class-label yes --negative-class-label no"
-
-        p, stdo, stde = _exec_shell_cmd(
-            cmd,
-            "Failed in {} command line! {}".format(ArgumentsOptions.MAIN_COMMAND, cmd),
-            assert_if_fail=False,
-        )
-
-        stdo_stde = str(stdo) + str(stde)
-
-        assert "Expected predictions to have columns ['yes', 'no']" in stdo_stde
-
     # testing negative cases: no artifact, no custom;
     @pytest.mark.parametrize(
         "framework, problem, language",
@@ -224,98 +183,6 @@ class TestOtherCases:
                 != -1
             )
             assert case
-
-    @pytest.mark.parametrize(
-        "framework, language, hooks_list, target_type",
-        [
-            (
-                None,
-                PYTHON_ALL_PREDICT_STRUCTURED_HOOKS,
-                CustomHooks.ALL_PREDICT_STRUCTURED,
-                REGRESSION,
-            ),
-            (None, R_ALL_PREDICT_STRUCTURED_HOOKS, CustomHooks.ALL_PREDICT_STRUCTURED, REGRESSION),
-            (
-                None,
-                R_ALL_PREDICT_STRUCTURED_HOOKS_LOWERCASE_R,
-                CustomHooks.ALL_PREDICT_STRUCTURED,
-                REGRESSION,
-            ),
-            (
-                None,
-                PYTHON_ALL_PREDICT_UNSTRUCTURED_HOOKS,
-                CustomHooks.ALL_PREDICT_UNSTRUCTURED,
-                UNSTRUCTURED,
-            ),
-            (
-                None,
-                R_ALL_PREDICT_UNSTRUCTURED_HOOKS,
-                CustomHooks.ALL_PREDICT_UNSTRUCTURED,
-                UNSTRUCTURED,
-            ),
-            (
-                None,
-                R_ALL_PREDICT_UNSTRUCTURED_HOOKS_LOWERCASE_R,
-                CustomHooks.ALL_PREDICT_UNSTRUCTURED,
-                UNSTRUCTURED,
-            ),
-        ],
-    )
-    def test_custom_model_with_all_hooks(
-        self, resources, framework, language, hooks_list, target_type, tmp_path,
-    ):
-        custom_model_dir = _create_custom_model_dir(resources, tmp_path, framework, None, language,)
-
-        input_dataset = resources.datasets(framework, REGRESSION)
-
-        output = tmp_path / "output"
-
-        cmd = "{} score --code-dir {} --input {} --output {} --target-type {}".format(
-            ArgumentsOptions.MAIN_COMMAND, custom_model_dir, input_dataset, output, target_type
-        )
-        _exec_shell_cmd(
-            cmd, "Failed in {} command line! {}".format(ArgumentsOptions.MAIN_COMMAND, cmd)
-        )
-        if hooks_list == CustomHooks.ALL_PREDICT_STRUCTURED:
-            preds = pd.read_csv(output)
-            assert all(val for val in (preds["Predictions"] == len(hooks_list)).values), preds
-        elif hooks_list == CustomHooks.ALL_PREDICT_UNSTRUCTURED:
-            with open(output) as f:
-                all_data = f.read()
-                assert str(len(hooks_list)) in all_data
-
-    @pytest.mark.parametrize(
-        "framework, language, hooks_list, target_type",
-        [
-            (None, R_INT_COLNAMES_BINARY, CustomHooks.SCORE, BINARY,),
-            (None, R_INT_COLNAMES_MULTICLASS, CustomHooks.SCORE, MULTICLASS,),
-        ],
-    )
-    @pytest.mark.parametrize("label_type", [int, float])
-    def test_custom_model_R_int_colnames_in_prediction_output(
-        self, resources, framework, language, hooks_list, target_type, label_type, tmp_path,
-    ):
-        custom_model_dir = _create_custom_model_dir(resources, tmp_path, framework, None, language,)
-        input_dataset = resources.datasets(framework, REGRESSION)
-        output = tmp_path / "output"
-
-        labels = [0, 1]
-        if target_type == MULTICLASS:
-            labels = [0, 1, 2]
-
-        labels = [label_type(l) for l in labels]
-
-        cmd = "{} score --code-dir {} --input {} --output {} --target-type {}".format(
-            ArgumentsOptions.MAIN_COMMAND, custom_model_dir, input_dataset, output, target_type
-        )
-        cmd = _cmd_add_class_labels(cmd, labels, target_type=target_type)
-
-        _exec_shell_cmd(
-            cmd, "Failed in {} command line! {}".format(ArgumentsOptions.MAIN_COMMAND, cmd)
-        )
-
-        preds = pd.read_csv(output)
-        assert all(preds.columns == [str(l) for l in labels])
 
     @pytest.mark.parametrize("language, language_suffix", [("python", ".py"), ("r", ".R")])
     def test_template_creation(self, language, language_suffix, tmp_path):
