+++ tests/fixtures/all_predict_structured_hooks_custom_lowercase_r.r
@@ -1,4 +1,4 @@
-prediction_value <- NaN
+prediction_value <<- NaN
 
 
 init <- function(...) {
