+++ tests/fixtures/all_predict_unstructured_hooks_custom.R
@@ -7,7 +7,7 @@
 # This is proprietary source code of DataRobot, Inc. and its affiliates.
 #
 # Released under the terms of DataRobot Tool and Utility Agreement.
-prediction_value <- NaN
+prediction_value <<- NaN
 
 init <- function(...) {
     #library(stringr)
